Update dpnet-IEnumHost patchset
- Add basic check for SessionName in EnumHost
- Fixed Connect, wrt to sending data
- Fixed SetClientInfo tests.
- Added dpnet-various-changes patch includes
  - Use the same socket for both EnumHost/Connect
  - Use DN_INTERNAL_MESSAGE_PLAYER_CONNECT_INFO_EX by default
  - Correct Name offset/size in Connect.
- Added dpnet/test
  - Show the EnumHost can be called multiple times (WIP)
    (needs to check the number of times MessageHandler is called).

From developing the above, It appear that Client uses a single
sock for listening for EnumHost/Connect and (binding between ports 2300-2400).
If not bound between these ports then Microsoft Message Analyser will
fail to see them as a DPNET packets.

A rewrite of how the packets are handle needs to be done. It's a basic
loop to collect packets, then sent to the correct MessageHandler(s).
Eg.  Enum Packer -> Enum Message Handler. and all other packets are
dropped if no handler is present.

A send queue will also need to be added and processed.