Updated msado15-Initial patchset
This is the way windows appears to Handle the fact we cannot
have multiple LOB's open at the same time.

We create an Accessor for each column of the table, and then
call GetData on each "Column".
The call to GetData appears to perform one of three actions.
SQLFetch (eg.  All bind columns are updated).
Nothing  (SQLFetch has been called, and nothing needs going).
SQLGetData (On non-bound columns, Large Object columns).

Currently this is WIP
- Memory leaks
- Structure should be used instead of multiple arrays.