Added odbc_dsn_valid patchset
This fixes the issue when adding a DSN, that it doesn't just write to
the system Registry point.  Since the SQLWrite/RemoveDSNFromIniW should
only be called from a setup DLL, it should have the proper config_mode
set.

PostgreSQL ODBC driver, calls though to SQLRemoveDSNFromIniW with the
previous name, which can be empty.  So wiping out any registry values under
the keys specified.