Added msvidc32_avi_format patchset
This isn't correct, and is completely in the wrong place.

The bug report has an uncompress msvc file of 24 bits which isn't supported by the
msvidc32 on windows. (or the codex spec for that matter).  Though appears to fix
the issue of the crash.

Under windows, a basic animation control, load a 24bit AVI doesn't appear to
load any msvid* dll's at all, which suggests that the image is handle internally
of the animation control found in comctrl32.